requirements.txt
spidev>=3.6
RPi.GPIO>=0.7.1
Pillow>=9.0.0
opencv-python-headless>=4.5.0
flask>=3.0
boto3>=1.28
rpi_ws281x>=5.0
requests[socks]>=2.28
pysocks>=1.7
openai>=1.0
httpx[socks]>=0.25
vosk>=0.3.45
